CafeteriaAPI/requirements.txt
alembic==1.7.7
aniso8601==9.0.1
click==8.1.3
colorama==0.4.4
Flask==2.1.2
Flask-JWT==0.3.2
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
passlib==1.7.4
PyJWT==1.4.2
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.36
thrift==0.16.0
Werkzeug==2.1.2